app/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
bcrypt==4.1.3
build==1.2.1
CacheControl==0.14.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cryptography==42.0.8
distlib==0.3.8
dnspython==2.6.1
dulwich==0.21.7
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.20.0
filelock==3.15.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2==3.1.4
keyring==24.3.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.3.0
msgpack==1.0.8
orjson==3.10.5
packaging==24.1
passlib==1.7.4
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
psycopg2==2.9.9
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.8.0
pydantic-extra-types==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.0
Pygments==2.18.0
PyJWT==2.8.0
pyproject_hooks==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rapidfuzz==3.9.4
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
SecretStorage==3.3.3
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
tomlkit==0.12.5
trove-classifiers==2024.7.2
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.26.3
watchfiles==0.22.0
websockets==12.0
